requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
qrcode==7.4.2
pillow==10.0.0
werkzeug==3.0.6
python-dotenv==1.0.1